Basic jump w/ jump and inair and land state + animations
A basic design of the jump, however, personally, hate the look of the mechanics behind it, mainly the animations for the inair state with gravity. There appears to be a node called AnimationTree (I think) where we can blend animations together, but since we're in too deep. just going to leave it to the side for now - gotta love refactoring