Fixing dumb bug + cleaning up states and transitions
Spent hours fixing a bug where there was some glitching between states; the states being the move and inair and land state. It's believed to be a problem with the physics engine regarding the function is on floor and the transitions being on the update function; it was solved with moving them to physics update - it's done now, no more thinky time. Some minor clean up, boring stuff..